move extra type info
